+++ storage/src/main/scala/io/casperlabs/storage/dag/FinalityStorage.scala
@@ -4,6 +4,6 @@ import io.casperlabs.storage.block.BlockStorage.BlockHash
 import simulacrum.typeclass
 
 @typeclass trait FinalityStorage[F[_]] {
-  def markAsFinalized(mainParent: BlockHash, secondary: Set[BlockHash], quorum: Int): F[Unit]
+  def markAsFinalized(mainParent: BlockHash, secondary: Set[BlockHash], quorum: BigInt): F[Unit]
   def isFinalized(block: BlockHash): F[Boolean]
 }
